+++ ios/CHANGELOG.md
@@ -27,6 +27,8 @@ Line wrap the file at 100 chars.                                              Th
 - Add search functionality to location selection view.
 - Wipe all settings on app reinstall.
 - Add a dedicated account button on the main view and remove it from settings.
+- Rotate public key from within packet tunnel when it detects that the key stored on backend does 
+  not match the one stored on device.
 
 
 ## [2023.2 - 2023-04-03]